# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grandhome Investments Real Estate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  name: string;
  address: string;
  /** Display price string (e.g. "$389,000" or "$2,200/mo") */
  price: string;
  /**
     * Numeric price for filtering
     * @nullable
     */
  priceValue?: number | null;
  beds: number;
  baths: number;
  sqft: number;
  /** House | Condo | Townhouse | Apartment */
  type: string;
  /** buy | rent */
  mode: string;
  /** New | Sale | Rent */
  tag: string;
  /** Array of base64 data URLs or image URLs */
  photos: string[];
  createdAt: string;
}

export interface PropertyInput {
  address: string;
  price: string;
  beds: number;
  baths: number;
  sqft: number;
  type: string;
  mode: string;
  photos?: string[];
}

export interface PropertyStats {
  totalListings: number;
  forSale: number;
  forRent: number;
  /** @nullable */
  avgPrice: number | null;
  newListings: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListPropertiesParams = {
/**
 * Minimum number of bedrooms
 */
beds?: number;
/**
 * Property type (House, Condo, Townhouse, Apartment)
 */
type?: string;
/**
 * Listing mode (buy, rent)
 */
mode?: string;
/**
 * Maximum price in thousands (e.g. 500 = $500K)
 */
maxPrice?: number;
/**
 * Search by address or name
 */
search?: string;
};