# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grandhome Investments Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  name: string;
  address: string;
  /** Display price string (e.g. "$389,000" or "$2,200/mo") */
  price: string;
  /**
     * Numeric price for filtering
     * @nullable
     */
  priceValue?: number | null;
  beds: number;
  baths: number;
  sqft: number;
  /** House | Condo | Townhouse | Apartment */
  type: string;
  /** buy | rent */
  mode: string;
  /** New | Sale | Rent */
  tag: string;
  /** Array of base64 data URLs or image URLs */
  photos: string[];
  createdAt: Date;
}