# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grandhome Investments Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyInput {
  address: string;
  price: string;
  beds: number;
  baths: number;
  sqft: number;
  type: string;
  mode: string;
  photos?: string[];
}