# lib/api-zod/src/generated/types/propertyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grandhome Investments Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyStats {
  totalListings: number;
  forSale: number;
  forRent: number;
  /** @nullable */
  avgPrice: number | null;
  newListings: number;
}